Rename exceptions filename to avoid camelCase
