+++ almanach/common/almanach_exception.py
