+++ almanach/core/controller.py
@@ -19,7 +19,7 @@ from datetime import timedelta
 from dateutil import parser as date_parser
 from pkg_resources import get_distribution
 
-from almanach.common.DateFormatException import DateFormatException
+from almanach.common.date_format_exception import DateFormatException
 from almanach.core.model import Instance, Volume, VolumeType
 from almanach import config
